Fix code duplication without introducing circular imports or breaking import rules
