+++ server/py/services/api/api/endpoints/nuclio.py
@@ -29,7 +29,6 @@ import mlrun.common.schemas
 import mlrun.common.schemas.model_monitoring.constants as mm_constants
 from mlrun.common.schemas.serving import DeployResponse
 from mlrun.config import config
-from mlrun.runtimes import RuntimeKinds
 from mlrun.utils import logger
 from mlrun.utils.helpers import generate_object_uri
 
@@ -45,6 +44,7 @@ import services.api.launcher
 from framework.api import deps
 from framework.constants import MINIMUM_CLIENT_VERSION_FOR_MM
 from services.api.crud.secrets import Secrets, SecretsClientType
+from services.api.utils.endpoints import start_model_endpoint_creation_background_task
 
 router = APIRouter()
 
@@ -218,61 +218,6 @@ async def delete_api_gateway(
             )
             return await client.delete_api_gateway(project_name=project, name=name)
         return await client.delete_api_gateway(project_name=project, name=name)
-
-
-# TODO: resolve duplication of this function without introducing circular imports
-async def start_model_endpoint_creation_background_task(
-    project: str,
-    name: str,
-    background_tasks: fastapi.BackgroundTasks,
-    function: dict,
-    db_session: sqlalchemy.orm.Session,
-    is_batch: bool,
-):
-    returned_background_tasks = mlrun.common.schemas.BackgroundTaskList(
-        background_tasks=[]
-    )
-    kind = function.get("kind")
-    if (
-        kind == RuntimeKinds.serving
-        or kind == RuntimeKinds.job
-        and function["spec"].get("serving_spec")
-    ):
-        monitoring_deployment = mm_deployment.MonitoringDeployment(project=project)
-        (
-            model_endpoints_instructions,
-            function,
-        ) = await monitoring_deployment._create_model_endpoints_instructions(
-            db_session=db_session,
-            function=function,
-            function_name=name,
-            project=project,
-            is_batch=is_batch,
-        )
-        logger.info(
-            "Creating Background Task for model endpoints creation",
-            project=project,
-            function=name,
-            is_batch=is_batch,
-        )
-        returned_background_task = await run_in_threadpool(
-            monitoring_deployment._create_model_endpoint_background_task,
-            db_session=db_session,
-            background_tasks=background_tasks,
-            project_name=project,
-            function_name=name,
-            function_tag=function.get("metadata", {}).get("tag") or "latest",
-            model_endpoints_instructions=model_endpoints_instructions,
-        )
-        returned_background_tasks.background_tasks.append(returned_background_task)
-
-    model_endpoint_creation_task_name = (
-        returned_background_tasks.background_tasks[0].metadata.name
-        if returned_background_tasks.background_tasks
-        else None
-    )
-
-    return function, model_endpoint_creation_task_name, returned_background_tasks
 
 
 @router.post(
